pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dftkit"
version = "0.1.0"
description = "Task-oriented CLI toolkit for DFT workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ase>=3.28.0",
    "click>=8.3.3",
    "click-option-group",
    "matplotlib>=3.10.9",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "platformdirs",
    "pydantic",
    "rich",
    "ruamel-yaml",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "seekpath>=2.2.1",
    "spglib>=2.7.0",
]

[project.scripts]
dftkit = "dftkit.cli:main"

[[tool.uv.index]]
url = "https://pypi.mirrors.ustc.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dftkit"]
